Refactor Device and Sensor classes to improve functionality and introduce EnergySensor. Update Device to handle multiple sensor types and enhance reading capabilities. Add vectorFind utility for vector operations

+++ server/SensorReading.hpp
@@ -1,6 +1,6 @@
 #pragma once
 struct SensorReading {
-    float value;
+    int value;
     float min_value;
     float max_value;
     double sum;
